20251226-epub翻译/requirements.txt
beautifulsoup4==4.12.3
ebooklib==0.18
lxml==5.3.0
python-docx==1.1.2
docxcompose==1.4.0
requests==2.32.3
tqdm==4.67.1
